// src/charts/src/sparkline/model/interface.ts
export interface ISparklineEventArgs {
    /** Defines the name of the event. */
    name: string;
    /** Defines the event cancel status. */
    cancel: boolean;
}
// ...
export interface ISparklineLoadedEventArgs extends ISparklineEventArgs {
    /** Defines the current sparkline instance. */
    sparkline: Sparkline;
}
// ...
export interface ISparklineResizeEventArgs {
    /** Defines the name of the Event. */
    name: string;
    /** Defines the previous size of the sparkline. */
    previousSize: Size;
    /** Defines the current size of the sparkline. */
    currentSize: Size;
    /** Defines the sparkline instance. */
    sparkline: Sparkline;
}